+++ test/compute-engine/latex-syntax/trigonometry.test.ts
@@ -196,3 +196,41 @@ describe('TRIGONOMETRIC FUNCTIONS', () => {
       N-mach    = 0.8090169943749475
     `));
 });
+
+describe('TRIGONOMETRIC DEGREES', () => {
+  test('\\cos(30\\deg)', () =>
+    expect(check('\\cos(30\\deg)')).toMatchInlineSnapshot(`
+      latex     = ["Cos", ["Degrees", 30]]
+      box       = ["Cos", ["Multiply", ["Rational", 1, 6], "Pi"]]
+      simplify  = ["Multiply", ["Rational", 1, 2], ["Sqrt", 3]]
+      N-auto    = 0.8660254037844386467637231707529361834714026269051903140279034897259665084544000185405730933786242878
+      N-mach    = 0.8660254037844387
+    `));
+
+  test('\\cos(30\\degree)', () =>
+    expect(check('\\cos(30\\degree)')).toMatchInlineSnapshot(`
+      latex     = ["Cos", ["Degrees", 30]]
+      box       = ["Cos", ["Multiply", ["Rational", 1, 6], "Pi"]]
+      simplify  = ["Multiply", ["Rational", 1, 2], ["Sqrt", 3]]
+      N-auto    = 0.8660254037844386467637231707529361834714026269051903140279034897259665084544000185405730933786242878
+      N-mach    = 0.8660254037844387
+    `));
+
+  test('\\cos(30^\\circ)', () =>
+    expect(check('\\cos(30^\\circ)')).toMatchInlineSnapshot(`
+      latex     = ["Cos", ["Degrees", 30]]
+      box       = ["Cos", ["Multiply", ["Rational", 1, 6], "Pi"]]
+      simplify  = ["Multiply", ["Rational", 1, 2], ["Sqrt", 3]]
+      N-auto    = 0.8660254037844386467637231707529361834714026269051903140279034897259665084544000185405730933786242878
+      N-mach    = 0.8660254037844387
+    `));
+
+  test('\\cos(\\ang{30})', () =>
+    expect(check('\\cos(\\ang{30})')).toMatchInlineSnapshot(`
+      latex     = ["Cos", ["Degrees", 30]]
+      box       = ["Cos", ["Multiply", ["Rational", 1, 6], "Pi"]]
+      simplify  = ["Multiply", ["Rational", 1, 2], ["Sqrt", 3]]
+      N-auto    = 0.8660254037844386467637231707529361834714026269051903140279034897259665084544000185405730933786242878
+      N-mach    = 0.8660254037844387
+    `));
+});