+++ test/compute-engine/smoke.test.ts
@@ -41,13 +41,13 @@ ce.assume('one', 1);
 // also should parse sub, i.e. f_{n-1} -> use sub as first params? (or last, as in log_2(x) -> log(x, 2))
 console.log(ce.parse("f'").json);
 
+const n1 = ce.parse('x_{1,2}');
+console.log(n1.toString());
+
 const expr = ce.parse('x^2').json;
 console.log(
   ce.box(['Integrate', expr, ['Element', 'x', ['Interval', 0, 1]]]).latex
 );
-
-const n1 = ce.parse('x_{1,2}');
-console.log(n1.toString());
 
 // Parsed as imaginary unit
 // -> don't have `i` (and `e`) in the dictionary mapping to `imaginaryUnit` and