+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ### Improvements
 
+- Support angles in degrees with `30\deg`, `30\degree`, `30^\circ` and
+  `\ang{30}`.
 - More accurate error expressions, for example if there is a missing closing
   delimiter an `["Error", ["ErrorCode", "'expected-closing-delimiter'", "')'"]]`
   is produced.